% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/harmonic_mean.R
\name{harmonic_mean}
\alias{harmonic_mean}
\title{Harmonic mean}
\usage{
harmonic_mean(x, na.rm = TRUE)
}
\arguments{
\item{x}{numeric vector}

\item{na.rm}{logical remove NAs prior or calculation}
}
\value{
harmonic mean of vector
}
\description{
Calculate the harmonic mean of a numeric vector
(will return NA if there are any negative numbers in the vector)
}
\examples{
data(nancycats)
pop.sizes <- table(pop(nancycats))
harmonic_mean(pop.sizes)
}

