% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\docType{package}
\name{mmod}
\alias{mmod}
\alias{mmod-package}
\alias{mmod-package}
\title{Modern Measures of Differentiation}
\description{
Population geneticists have traditionally used Nei's Gst (often confusingly called 
Fst...) to measure divergence between populations. Recently, it has become 
clear that simple intereptations of the value of Gst can be misleading. For
this reason several new measures differntiation have been developed. mmod is
a package that brings some of these measures to R.
}
\details{
The vignette for this package ( avaliable using 
vignette("demo", package="mmod") from within R) contains an introduction to
these methods and and example usage for this package. I strongly suggest new
users start by reading this documentation.
}
